Split clickhouse integration tests into separate files for each method (#11209)
* separate clickhouse tests into different files

* change name of cna counts test file

---------
