+++ src/test/java/org/cbioportal/persistence/mybatisclickhouse/ClinicalEventTypeCountsTest.java
@@ -1,0 +1,67 @@
+package org.cbioportal.persistence.mybatisclickhouse;
+
+import org.cbioportal.persistence.helper.StudyViewFilterHelper;
+import org.cbioportal.persistence.mybatisclickhouse.config.MyBatisConfig;
+import org.cbioportal.web.parameter.DataFilter;
+import org.cbioportal.web.parameter.DataFilterValue;
+import org.cbioportal.web.parameter.StudyViewFilter;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+@RunWith(SpringRunner.class)
+@Import(MyBatisConfig.class)
+@DataJpaTest
+@DirtiesContext
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+@ContextConfiguration(initializers = AbstractTestcontainers.Initializer.class)
+public class ClinicalEventTypeCountsTest extends AbstractTestcontainers {
+
+    private static final String STUDY_TCGA_PUB = "study_tcga_pub";
+
+    @Autowired
+    private StudyViewMapper studyViewMapper;
+
+    @Test
+    public void getClinicalEventTypeCounts() {
+        StudyViewFilter studyViewFilter = new StudyViewFilter();
+        studyViewFilter.setStudyIds(List.of(STUDY_TCGA_PUB));
+
+        var clinicalEventTypeCounts = studyViewMapper.getClinicalEventTypeCounts(StudyViewFilterHelper.build(studyViewFilter, null, null));
+
+        assertEquals(4, clinicalEventTypeCounts.size());
+
+        var clinicalEventTypeCountOptional = clinicalEventTypeCounts.stream().filter(ce -> ce.getEventType().equals("Treatment"))
+            .findFirst();
+
+        assertTrue(clinicalEventTypeCountOptional.isPresent());
+        assertEquals(1, clinicalEventTypeCountOptional.get().getCount().intValue());
+
+        DataFilter dataFilter = new DataFilter();
+        DataFilterValue dataFilterValue = new DataFilterValue();
+        dataFilterValue.setValue("Treatment");
+        dataFilter.setValues(List.of(dataFilterValue));
+        studyViewFilter.setClinicalEventFilters(List.of(dataFilter));
+
+        clinicalEventTypeCounts = studyViewMapper.getClinicalEventTypeCounts(StudyViewFilterHelper.build(studyViewFilter, null, null));
+
+        assertEquals(3, clinicalEventTypeCounts.size());
+
+        clinicalEventTypeCountOptional = clinicalEventTypeCounts.stream().filter(ce -> ce.getEventType().equals("status"))
+            .findFirst();
+
+        assertFalse(clinicalEventTypeCountOptional.isPresent());
+    }
+}
