+++ src/test/java/org/cbioportal/persistence/mybatisclickhouse/SampleTreatmentCountsTest.java
@@ -1,0 +1,71 @@
+package org.cbioportal.persistence.mybatisclickhouse;
+
+import org.cbioportal.model.TemporalRelation;
+import org.cbioportal.persistence.helper.StudyViewFilterHelper;
+import org.cbioportal.persistence.mybatisclickhouse.config.MyBatisConfig;
+import org.cbioportal.web.parameter.StudyViewFilter;
+import org.cbioportal.web.parameter.filter.AndedSampleTreatmentFilters;
+import org.cbioportal.web.parameter.filter.OredSampleTreatmentFilters;
+import org.cbioportal.web.parameter.filter.SampleTreatmentFilter;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.ContextConfiguration;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import java.util.List;
+
+import static org.junit.Assert.assertEquals;
+
+@RunWith(SpringRunner.class)
+@Import(MyBatisConfig.class)
+@DataJpaTest
+@DirtiesContext
+@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE)
+@ContextConfiguration(initializers = AbstractTestcontainers.Initializer.class)
+public class SampleTreatmentCountsTest extends AbstractTestcontainers {
+
+    private static final String STUDY_TCGA_PUB = "study_tcga_pub";
+
+    @Autowired
+    private StudyViewMapper studyViewMapper;
+
+    @Test
+    public void getSampleTreatmentCounts() {
+        StudyViewFilter studyViewFilter = new StudyViewFilter();
+        studyViewFilter.setStudyIds(List.of(STUDY_TCGA_PUB));
+
+
+        var totalSampleTreatmentCount = studyViewMapper.getTotalSampleTreatmentCounts(StudyViewFilterHelper.build(studyViewFilter, null, null));
+
+        var sampleTreatmentCounts = studyViewMapper.getSampleTreatmentCounts(StudyViewFilterHelper.build(studyViewFilter, null, null));
+
+        assertEquals(1, totalSampleTreatmentCount);
+        assertEquals("madeupanib", sampleTreatmentCounts.getFirst().treatment());
+        assertEquals(1, sampleTreatmentCounts.getFirst().postSampleCount());
+        assertEquals(0, sampleTreatmentCounts.getFirst().preSampleCount());
+
+        SampleTreatmentFilter filter = new SampleTreatmentFilter();
+        filter.setTreatment("madeupanib");
+        filter.setTime(TemporalRelation.Pre);
+
+        OredSampleTreatmentFilters oredSampleTreatmentFilters = new OredSampleTreatmentFilters();
+        oredSampleTreatmentFilters.setFilters(List.of(filter));
+
+        AndedSampleTreatmentFilters andedSampleTreatmentFilters = new AndedSampleTreatmentFilters();
+        andedSampleTreatmentFilters.setFilters(List.of(oredSampleTreatmentFilters));
+        studyViewFilter.setSampleTreatmentFilters(andedSampleTreatmentFilters);
+
+        totalSampleTreatmentCount = studyViewMapper.getTotalSampleTreatmentCounts(StudyViewFilterHelper.build(studyViewFilter, null, null));
+
+        sampleTreatmentCounts = studyViewMapper.getSampleTreatmentCounts(StudyViewFilterHelper.build(studyViewFilter, null, null));
+
+        assertEquals(0, totalSampleTreatmentCount);
+        assertEquals(0, sampleTreatmentCounts.size());
+
+    }
+}
